+++ packages/hash/playwright/tests/page-creation.spec.ts
@@ -1,8 +1,7 @@
-import { test, expect } from "@playwright/test";
 import { sleep } from "@hashintel/hash-shared/sleep";
+import { test, expect } from "./shared/runtime";
 import { resetDb } from "./shared/reset-db";
 import { loginUsingTempForm } from "./shared/login-using-temp-form";
-import { failOnConsoleOutput } from "./shared/console-checks";
 
 const pageNameSuffix = Date.now();
 const pageNameFallback = "Untitled";
@@ -14,9 +13,8 @@ const placeholderSelector =
   "text=Type / to browse blocks, or @ to browse entities";
 const modifierKey = process.platform === "darwin" ? "Meta" : "Control";
 
-test.beforeEach(async ({ page }) => {
+test.beforeEach(async () => {
   await resetDb();
-  failOnConsoleOutput(page);
 });
 
 test("user can create page", async ({ page }) => {
